refactor: 접근제어자 및 패키지 변경
