+++ src/test/java/lotto/domain/LottoNumbersTest.java
@@ -1,10 +1,9 @@
-package lotto.domain.number;
+package lotto.domain;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.List;
 import java.util.stream.Stream;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
